% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spicy-package.R
\docType{package}
\name{spicy-package}
\alias{spicy}
\alias{spicy-package}
\title{spicy: Descriptive Statistics, Summary Tables, and Data Management Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools for descriptive data analysis, variable inspection, data management, and tabulation workflows in 'R'. Summarizes variable metadata, labels, classes, missing values, and representative values, with support for readable frequency tables, cross-tabulations, association measures for contingency tables (Cramer's V, Phi, Goodman-Kruskal Gamma, Kendall's Tau-b, Somers' D, and others), categorical and continuous summary tables, including APA-style reporting outputs. Includes helpers for interactive codebooks, variable label extraction, clipboard export, and row-wise descriptive summaries. Designed to make descriptive analysis faster, clearer, and easier to work with in practice.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amaltawfik/spicy/}
  \item \url{https://amaltawfik.github.io/spicy/}
  \item Report bugs at \url{https://github.com/amaltawfik/spicy/issues}
}

}
\author{
\strong{Maintainer}: Amal Tawfik \email{amal.tawfik@hesav.ch} (\href{https://orcid.org/0009-0006-2422-1555}{ORCID}) (\href{https://ror.org/04j47fz63}{ROR}) [copyright holder]

}
\keyword{internal}
